Added: router for each menu and theme switcher
Now use can switch btw themes using a small button on the bottom side of the sidebar.
When clicking a menu, user will get redirected to the current slug and the menu item will be focused with an overlay color to indicate where is the user actually.
Modified some aspects colors to work with theme switcher.
Added routes for every menu as an example and for testing porpuses.
Added user data from menus.ts to the sidebar header.
To much todo by now, but i still working on it :).